backend/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.8",
    "google-generativeai>=0.8.4",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=3.4.1",
]
